refactor(core): keep only Transform as a compiler class
Roll Parse and Generate logic back into KubbDriver as private methods
(#parseInput, #runGenerators). Transform stays as a small stateful
class at the top level — it is the one piece with state worth its own
type. The compiler subfolder goes away.

The renderer-or-file dispatch (previously Generate.apply, originally
applyHookResult) becomes KubbDriver.applyResult — a static so mocks can
call it without instantiating a driver. The "Hook" qualifier dropped
since it handles generator results, not hook results specifically.

End state:
- packages/core/src/Transform.ts (registry)
- packages/core/src/KubbDriver.ts (everything else)
- No GenerateHost interface, no Parse/Generate class wrappers, no
  threading host or fileManager through static method params.

KubbDriver test gains five direct tests for applyResult covering the
array, renderer-stream, renderer-async, and no-op paths.
