+++ packages/core/src/KubbDriver.test.ts
@@ -1,6 +1,9 @@
 import { AsyncEventEmitter } from '@internals/utils'
+import { createFile } from '@kubb/ast'
+import type { FileNode } from '@kubb/ast'
 import { createMockedAdapter } from '@kubb/core/mocks'
-import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
+import { afterEach, beforeEach, describe, expect, it, test, vi } from 'vitest'
+import { FileManager } from './FileManager.ts'
 import { KubbDriver } from './KubbDriver.ts'
 import type { Config, KubbHooks, Middleware, Plugin } from './types.ts'
 import { fsStorage } from './storages/fsStorage.ts'
@@ -132,5 +135,71 @@ describe('PluginDriver', () => {
     })
 
     expect(external).toHaveBeenCalledTimes(1)
+  })
+})
+
+function file(name: string): FileNode {
+  return createFile({ baseName: `${name}.ts`, path: `${name}.ts` })
+}
+
+describe('KubbDriver.applyResult', () => {
+  it('does nothing on null or undefined', () => {
+    const fileManager = new FileManager()
+    const upsert = vi.spyOn(fileManager, 'upsert')
+
+    KubbDriver.applyResult({ result: null, fileManager })
+    KubbDriver.applyResult({ result: undefined, fileManager })
+
+    expect(upsert).not.toHaveBeenCalled()
+  })
+
+  it('upserts every file when the result is an Array<FileNode>', () => {
+    const fileManager = new FileManager()
+    const files = [file('a'), file('b')]
+
+    KubbDriver.applyResult({ result: files, fileManager })
+
+    expect(fileManager.files.map((f) => f.name)).toStrictEqual(['a', 'b'])
+  })
+
+  it('ignores non-array results when no rendererFactory is provided', () => {
+    const fileManager = new FileManager()
+    const upsert = vi.spyOn(fileManager, 'upsert')
+
+    KubbDriver.applyResult({ result: { kind: 'element' }, fileManager })
+
+    expect(upsert).not.toHaveBeenCalled()
+  })
+
+  it('routes element results through the rendererFactory stream when present', () => {
+    const fileManager = new FileManager()
+    const rendered = [file('rendered-1'), file('rendered-2')]
+    const rendererFactory = vi.fn(() => ({
+      stream: vi.fn(function* () {
+        yield* rendered
+      }),
+      [Symbol.dispose]: () => {},
+    }))
+
+    KubbDriver.applyResult({ result: { kind: 'element' }, fileManager, rendererFactory: rendererFactory as never })
+
+    expect(rendererFactory).toHaveBeenCalledOnce()
+    expect(fileManager.files.map((f) => f.name)).toStrictEqual(['rendered-1', 'rendered-2'])
+  })
+
+  it('routes element results through the async render path when no stream is exposed', async () => {
+    const fileManager = new FileManager()
+    const renderer = {
+      render: vi.fn(async () => {}),
+      files: [file('async-1')],
+      [Symbol.dispose]: () => {},
+    }
+    const rendererFactory = vi.fn(() => renderer)
+
+    const result = KubbDriver.applyResult({ result: { kind: 'element' }, fileManager, rendererFactory: rendererFactory as never })
+    await result
+
+    expect(renderer.render).toHaveBeenCalledOnce()
+    expect(fileManager.files.map((f) => f.name)).toStrictEqual(['async-1'])
   })
 })
