+++ packages/core/src/mocks.ts
@@ -2,7 +2,6 @@ import { resolve } from 'node:path'
 import type { FileNode, InputMeta, OperationNode, SchemaNode, Visitor } from '@kubb/ast'
 import { transform } from '@kubb/ast'
 import { FileManager } from './FileManager.ts'
-import { Generate } from './compiler/Generate.ts'
 import { KubbDriver } from './KubbDriver.ts'
 import type { Adapter, AdapterFactoryOptions, Config, Generator, GeneratorContext, NormalizedPlugin, PluginFactoryOptions } from './types.ts'
 
@@ -124,7 +123,7 @@ export async function renderGeneratorSchema<TOptions extends PluginFactoryOption
     ...context,
     options: opts.options,
   })
-  await Generate.apply({ result, fileManager: opts.driver.fileManager, rendererFactory: generator.renderer })
+  await KubbDriver.applyResult({ result, fileManager: opts.driver.fileManager, rendererFactory: generator.renderer })
 }
 
 /**
@@ -148,7 +147,7 @@ export async function renderGeneratorOperation<TOptions extends PluginFactoryOpt
     ...context,
     options: opts.options,
   })
-  await Generate.apply({ result, fileManager: opts.driver.fileManager, rendererFactory: generator.renderer })
+  await KubbDriver.applyResult({ result, fileManager: opts.driver.fileManager, rendererFactory: generator.renderer })
 }
 
 /**
@@ -172,5 +171,5 @@ export async function renderGeneratorOperations<TOptions extends PluginFactoryOp
     ...context,
     options: opts.options,
   })
-  await Generate.apply({ result, fileManager: opts.driver.fileManager, rendererFactory: generator.renderer })
+  await KubbDriver.applyResult({ result, fileManager: opts.driver.fileManager, rendererFactory: generator.renderer })
 }
